Quadratic curve is applied in pseudo_sky.cc, etc.:  - Applied quadratic curve (pseudo_sky.cc).  - Supported 32-bit float tiff (make_flat.cc, proc_images.cc, stack_images.cc).  - ISO speed, shutter, aperture, focal length are ebmedded in tiff files (raw2tiff.c).

+++ src/read_tiff24or48_to_float.h
@@ -1,4 +1,4 @@
 
 static int read_tiff24or48_to_float( const char *filename_in,
-    mdarray_float *ret_img_buf, mdarray_uchar *ret_icc_buf, size_t *ret_bytes,
+    mdarray_float *ret_img_buf, mdarray_uchar *ret_icc_buf, int *ret_bytes,
     float camera_calibration1_ret[] );
